Update to support TFS test case query files
Added support to retrieve personal and team TFS queries and filter them
to make queries that contain "Test case" and "Test Case Category"
options.

To support this I added a Select Query button that opens a dialog that
displays the available queries in one of two formats, a tree view
(similar to what TFS offers) and a flat list.  The user can switch
between the two by selecting a radio button.

Once a query is selected, it's executed and a new window opens that
lists the test cases from the query.  Clicking on a test case entry in
the window will cause the Test Steps Editor to automatically load the
test cases.

Finally, I added a "Refresh" button that will get the test case ID of
the currently selected test case in the editor, close the tab and
re-retrieve the test case steps from the server.